move chain.py into submodule, split out chain data
